LayoutTests/imported/w3c/web-platform-tests/shadow-dom/Extensions-to-Event-Interface-expected.txt


PASS composedPath() must exist on Event 
PASS composedPath() must return an empty array when the event has not been dispatched 
PASS composedPath() must return an empty array when the event is no longer dispatched 
PASS composed must exist on Event 
PASS composed on EventInit must default to false 
PASS composed on EventInit must set the composed flag 
FAIL The event must propagate out of open mode shadow boundaries when the composed flag is set assert_array_equals: value is 5, expected array
FAIL The event must propagate out of closed mode shadow boundaries when the composed flag is set assert_array_equals: value is 5, expected array
FAIL The event must not propagate out of open mode shadow boundaries when the composed flag is unset assert_array_equals: value is 2, expected array
FAIL The event must not propagate out of closed mode shadow boundaries when the composed flag is unset assert_array_equals: value is 2, expected array
FAIL The event must not propagate out of open mode shadow boundaries when the composed flag is unset on an event with relatedTarget assert_array_equals: value is 2, expected array
FAIL The event must not propagate out of closed mode shadow boundaries when the composed flag is unset on an event with relatedTarget assert_array_equals: value is 2, expected array
FAIL The event must not propagate out of open mode shadow tree of the target but must propagate out of inner shadow trees when the scoped flag is set assert_array_equals: value is 5, expected array
FAIL The event must not propagate out of closed mode shadow tree of the target but must propagate out of inner shadow trees when the scoped flag is set assert_array_equals: value is 5, expected array
FAIL The event must propagate out of open mode shadow tree in which the relative target and the relative related target are the same assert_array_equals: value is 4, expected array
FAIL The event must propagate out of closed mode shadow tree in which the relative target and the relative related target are the same assert_array_equals: value is 4, expected array
FAIL composedPath() must contain and only contain the unclosed nodes of target in open mode shadow trees assert_array_equals: value is 9, expected array
FAIL composedPath() must contain and only contain the unclosed nodes of target in closed mode shadow trees assert_array_equals: value is 9, expected array
